DISPLAY-990: Moved commands around
