+++ src/Command/Users/AddUserCommand.php
@@ -9,7 +9,7 @@
  * file that was distributed with this source code.
  */
 
-namespace App\Command;
+namespace App\Command\Users;
 
 use App\Entity\User;
 use App\Entity\UserRoleTenant;
@@ -27,7 +27,6 @@ use Symfony\Component\Console\Output\OutputInterface;
 use Symfony\Component\Console\Style\SymfonyStyle;
 use Symfony\Component\PasswordHasher\Hasher\UserPasswordHasherInterface;
 use Symfony\Component\Stopwatch\Stopwatch;
-
 use function Symfony\Component\String\u;
 
 /**
